refactor: extract UI handlers
